Query parameters selection and reset
For now, only range year start and range year end. And it's already a little bit hard to understand what's going on. I'd like to also add a chloropleth of a selected genre across cantons, and perhaps the same for all other relevant categories (stores, states, and platforms), but I will have to make it clearer what does what and what data is shown where, and how to select the data the user wants